adding optimization of doubles to floats, mem down again, need SIMD for performance

+++ vsm/model/_cgs_update.pyx
@@ -93,9 +93,9 @@ def cgs_update_int_char(int itr,
 @cython.cdivision(True)
 def cgs_update_int_short(int itr, 
                unsigned int [:] corpus,
-               double [:,:] word_top,
-               double [:] inv_top_sums,
-               double [:,:] top_doc,
+               float [:,:] word_top,
+               float [:] inv_top_sums,
+               float [:,:] top_doc,
                unsigned short [:] Z,
                int [:] indices,
                str mtrand_str,
@@ -108,22 +108,22 @@ def cgs_update_int_short(int itr,
     cdef int N = indices.shape[0]
     cdef int K = word_top.shape[1]
 
-    cdef double log_p = 0
-    cdef double [:,:] log_wk = np.log(np.asarray(word_top) * 
+    cdef float log_p = 0
+    cdef float [:,:] log_wk = np.log(np.asarray(word_top) * 
                                       np.asarray(inv_top_sums))
-    cdef double [:,:] log_kd = np.log(np.asarray(top_doc) /
+    cdef float [:,:] log_kd = np.log(np.asarray(top_doc) /
                                       np.asarray(top_doc).sum(0))
 
     cdef object np_random_state = np.random.RandomState()
     np_random_state.set_state((mtrand_str, mtrand_keys, 
                                mtrand_pos, mtrand_has_gauss, 
                                mtrand_cached_gaussian))
-    cdef double [:] samples = np_random_state.uniform(size=V)
+    cdef float [:] samples = np_random_state.uniform(size=V).astype(np.float32)
     cdef object mtrand_state = np_random_state.get_state()
-    cdef double [:] dist = np.zeros((K,), dtype=np.float64)
-    cdef double [:] cum_dist = np.zeros((K,), dtype=np.float64)
+    cdef float [:] dist = np.zeros((K,), dtype=np.float32)
+    cdef float [:] cum_dist = np.zeros((K,), dtype=np.float32)
 
-    cdef double r, s
+    cdef float r, s
     cdef long start, stop, doc_len, offset
     cdef unsigned short k, t
     cdef Py_ssize_t i, j, idx, w
@@ -144,6 +144,7 @@ def cgs_update_int_short(int itr,
             idx = offset + j
             w,k = corpus[idx], Z[idx]
 
+            # float = float, float
             log_p += log_wk[w, k] + log_kd[k, i]
 
             if itr > 0:
@@ -153,11 +154,13 @@ def cgs_update_int_short(int itr,
                 top_doc[k, i] -= 1
 
             for t in range(K):
+                #float =  float, float, float
                 dist[t] = inv_top_sums[t] * word_top[w,t] * top_doc[t,i]
                 if t==0:
                     cum_dist[t] = dist[t]
                 else:
                     cum_dist[t] = cum_dist[t-1] + dist[t]
+            # float = float, float
             r = samples[idx] * cum_dist[K-1]
             for k in range(K):
                 if r < cum_dist[k]:
@@ -180,9 +183,9 @@ def cgs_update_int_short(int itr,
 @cython.cdivision(True)
 def cgs_update_short_char(int itr, 
                unsigned short [:] corpus,
-               double [:,:] word_top,
-               double [:] inv_top_sums,
-               double [:,:] top_doc,
+               float [:,:] word_top,
+               float [:] inv_top_sums,
+               float [:,:] top_doc,
                unsigned char [:] Z,
                int [:] indices,
                str mtrand_str,
@@ -195,22 +198,22 @@ def cgs_update_short_char(int itr,
     cdef int N = indices.shape[0]
     cdef int K = word_top.shape[1]
 
-    cdef double log_p = 0
-    cdef double [:,:] log_wk = np.log(np.asarray(word_top) * 
+    cdef float log_p = 0
+    cdef float [:,:] log_wk = np.log(np.asarray(word_top) * 
                                       np.asarray(inv_top_sums))
-    cdef double [:,:] log_kd = np.log(np.asarray(top_doc) /
+    cdef float [:,:] log_kd = np.log(np.asarray(top_doc) /
                                       np.asarray(top_doc).sum(0))
 
     cdef object np_random_state = np.random.RandomState()
     np_random_state.set_state((mtrand_str, mtrand_keys, 
                                mtrand_pos, mtrand_has_gauss, 
                                mtrand_cached_gaussian))
-    cdef double [:] samples = np_random_state.uniform(size=V)
+    cdef float [:] samples = np_random_state.uniform(size=V).astype(np.float32)
     cdef object mtrand_state = np_random_state.get_state()
-    cdef double [:] dist = np.zeros((K,), dtype=np.float64)
-    cdef double [:] cum_dist = np.zeros((K,), dtype=np.float64)
+    cdef float [:] dist = np.zeros((K,), dtype=np.float32)
+    cdef float [:] cum_dist = np.zeros((K,), dtype=np.float32)
 
-    cdef double r, s
+    cdef float r, s
     cdef long start, stop, doc_len, offset
     cdef unsigned char k,t
     cdef Py_ssize_t i, j, idx, w
